requirements.txt
# Core dependencies
click>=8.0.0
flask>=2.3.0
requests>=2.28.0
Pillow>=10.1.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional - uncomment as needed
# python-dotenv>=1.0.0
# gunicorn>=20.1.0
# flask-cors>=4.0.0 